UPDATE: Handle Supplier Order Email Accept Button
This commit adds an html email template that will be sent to the supplier
when an order is placed. The email contains an accept button that the
supplier can click to accept the order. The button will redirect the
supplier to the supplier order page where they can view the order
details and accept the order. This page will show the order status
if it is anything other than PENDING.

Two new apis are added to the api/views.py file.
- api/order/<uuid:order_id>/view/ This renders the order page for the
  supplier to view the order details and accept the order.
- api/order/<uuid:order_id>/accept/ This changes the order status to SCHEDULED.

Also added an encrypt and decrypt function to the api/utils/utils.py.
This can be used to encrypt and decrypt any string using a secret key.
This is used to encrypt the order id in the email link to ensure
that any calls to the above apis are from our server.