UPDATE: Add User Invite policy check on UserInvite model save.
This commits adds a check to UserInvite save to ensure that the user
inviting is allowed to invite users to the group. This is done by checking
the UserGroup's policy for inviting users. Admins are allowed to invite
users to the group. If a user is allowed to be added, then the
UserInvite model is not actually saved, but the user is added to the group.