FIX: Do not set PayMent inactive on BasisTheory Payment Error
There seems to be inconsistency on a BasisTheory PaymentRequired error,
sometimes it shows and other times it doesn't.
This commit ensures that the payment method is not set to inactive
when a BasisTheory 202 PaymentRequired error is raised.
Also the log is turned off to avoid over logging.
This is still printed to console and so can be seen in
the digital ocean logs.