Update: Limit who can add new users in supplier portal
If a seller is not selected, then only allow superusers
to add new users, because this adds to the Downstream team.
If seller is selected, then only allow users who are staff
or are of type admin. If no seller is selected, then show popup
asking user to select a seller.