UPDATE: Automatically redirect on login and post login
This commit updates a new feature to the login router that will automatically
redirect the user to Auth0 login page instead of first
showing the django login page. This still grabs the `next`
query param and sets in the session so that post login the user is
redirected to the correct page.

This commit also updates the feature where if the user's
redirect_url is set, then the email they are sent first
routes through our server before automatically redirecting
to the change password url. This is done so that a session variable
can be set. Auth0 Tenant Login URI points to our server on post
login so that we can first check for that session variable
in order to redirect to the correct page.