UPDATE: Supplier Chat Header Message Dropdown
Add a message dropdown to the header and show the
open conversations in the dropdown, the last message.
Add a badge to the header to show the number of open conversations.
Add the ability to read and respond to non order based
conversations in the chat. This could happen if an admin
starts a conversation with a supplier.