UPDATE: Make Add New User on Company page work
This opens a modal and on confirmation will impersonate
the company and then open the correct page.

I also added code to make this process all local in the modal,
but it is a bit more complex. I believe it is a working
solution, so we could use it if we want to.