FIX: Ensure all emails are non empty in supplier approval emails
This ensures that all emails are non empty in supplier approval emails.
This is to prevent the case where an empty email will cause
the EmailNotification system to error due to trying to add an
empty email address, which then means the order approval
email doesn't ever send, even to dispatch. The fix will catch
the empty email case and will log an error as well as
send the email to the dispatch email address
(add_internal_email_to_queue).