FIX: Handle add payment method fo staff users not currently impersonating.
If a staff user is not impersonating a customer,
then add payment api will use the passed in user_address
to get the desired customer. If the staff user is
impersonating a customer, or if it is the actual
customer, then the user and unser_group will be
already set correctly.