FIX: Update to use schedule_window instead of time_slot
This will use the Order/transaction schedule_window instead of the
time_slot for the order confirmation email and supplier email templates.
It will always fall back to the time_slot if the schedule_window
is not set. If both are not set, it will show a default message "Any".
This closes #733