FIX: Implement Post/Redirect/Get on new Order to fix page reload POSTs
This will fix the issue where when directly after an
item is added to the cart, then the user refreshes the
page, the item is added again. This is because the page
is reloaded with the POST request still in the browser
history and the browser will re-submit the POST request
when the page is reloaded. https://en.wikipedia.org/wiki/Post/Redirect/Get