FIX: Fix terms agreement
Fix when and where the terms agreement is shown to the user.
If the parent Booking has never had a signed terms agreement,
the user will be prompted to sign the terms agreement
when they checkout, even if the current order is a SWAP,
which typically doesn't require a signature.