FIX: Fix redirect loop on non staff login on index
The scenario was that when a non staff user tried to login
on the base url, it set the url to /admin/, which only allows staff,
the user was then redirected to the login page since they did
not have access to /admin/. This created a redirect loop.
The fix was to redirect the user to the customer home page
if they are not staff and they were attempting to reach /admin/.
The second fix here was to check if they were attempting
to reach /supplier/ but were not a supplier, in which case
they would be redirected to the customer home page.