UPDATE: Duplicate accounts Auth0.
On Auth0 signup, users can enter a same Company name
that already exists and think they are added to that company,
but instead they are part of a new company with the same name.
This commit attempts to mitigate this by checking
if their email domain is already associated with a company.