FIX: Fix page load error
This happened because seller admin modified the base class
read_only list, which was itself a reference to audit_fields,
which were sometimes used in non read only admin pages.
This caused an error because the promary key id is not allowed
to be modified. Example to reproduce the error:
1. Go to the seller admin page and click on any seller
2. Go to MainProduct page and click on any main product.